billing: Remove invoice option for users with fixed price plan offer.
Currently, the user selected invoice option requires manual license
management, and a fixed-price plan offer requires automatic license
management.

If we're offering a fixed-price plan and the user wants to pay by
invoice, then the current system in place require us to create the
invoice in Stripe first and then input it manually via the support
page for the customer.