docs: Add WSL rebuild steps.
Previously, the rebuild steps for WSL in the recommended setup guide
incorrectly just included the Vagrant instructions.

This commit fixes the issue by adding the appropriate steps to delete
the WSL instance using `wsl --unregister` in the docs.

Additionally, it includes steps to rebuild the development database,
which is a much faster alternative for folks who just want that.

Fixes #32402.