support: Pass remote_support_view boolean in context for views.
Many of the templates for the remote and Cloud support views are
designed to be shared, but some features/actions have not yet been
implemented for both types of support views.

Adds a remote_support_view boolean to each type of support view
context so that it's easy to add checks to existing templates.

Updates the Cloud support view to use the current_plan_forms_support
template, and fixes existing bug of offering the upgrade to Plus
option for modifying a Cloud plan when the current plan is a
fixed-price plan.

Prep for implementing fixed-price plan offers in Cloud support view.