topics: Fix topic order after moving message without automated notice.
Previously, when a message was moved between two topics without
sending an automated notice, the ordering of these topics in the
left sidebar was not updated.

This is because the updated `max_message_id` of
`stream_topic_history` was always computed before the moved
message object's topic/channel were updated. This caused the
`max_message_id` of `stream_topic_history` which determines
the ordering of topics to always be the same.

This is fixed by making sure that the channel/topics of the objects
of the messages which were supposed to be moved were updated
before we compute this `max_message_id`

Co-authored-by: Rohan Gudimetla <rohan.gudimetla07@gmail.com>